+++ crates/native/src/bridge_generated.rs
@@ -683,7 +683,7 @@ fn wire_register_track_observer_impl(
         },
     )
 }
-fn wire_set_volume_observer_enabled_impl(
+fn wire_set_audio_level_observer_enabled_impl(
     port_: MessagePort,
     track_id: impl Wire2Api<String> + UnwindSafe,
     peer_id: impl Wire2Api<Option<u64>> + UnwindSafe,
@@ -691,7 +691,7 @@ fn wire_set_volume_observer_enabled_impl(
 ) {
     FLUTTER_RUST_BRIDGE_HANDLER.wrap::<_, _, _, (), _>(
         WrapInfo {
-            debug_name: "set_volume_observer_enabled",
+            debug_name: "set_audio_level_observer_enabled",
             port: Some(port_),
             mode: FfiCallMode::Normal,
         },
@@ -699,7 +699,9 @@ fn wire_set_volume_observer_enabled_impl(
             let api_track_id = track_id.wire2api();
             let api_peer_id = peer_id.wire2api();
             let api_enabled = enabled.wire2api();
-            move |task_callback| set_volume_observer_enabled(api_track_id, api_peer_id, api_enabled)
+            move |task_callback| {
+                set_audio_level_observer_enabled(api_track_id, api_peer_id, api_enabled)
+            }
         },
     )
 }
@@ -1656,7 +1658,9 @@ impl support::IntoDart for TrackEvent {
     fn into_dart(self) -> support::DartAbi {
         match self {
             Self::Ended => vec![0.into_dart()],
-            Self::VolumeUpdated(field0) => vec![1.into_dart(), field0.into_into_dart().into_dart()],
+            Self::AudioLevelUpdated(field0) => {
+                vec![1.into_dart(), field0.into_into_dart().into_dart()]
+            }
         }
         .into_dart()
     }
@@ -2029,13 +2033,13 @@ mod io {
     }
 
     #[no_mangle]
-    pub extern "C" fn wire_set_volume_observer_enabled(
+    pub extern "C" fn wire_set_audio_level_observer_enabled(
         port_: i64,
         track_id: *mut wire_uint_8_list,
         peer_id: *mut u64,
         enabled: bool,
     ) {
-        wire_set_volume_observer_enabled_impl(port_, track_id, peer_id, enabled)
+        wire_set_audio_level_observer_enabled_impl(port_, track_id, peer_id, enabled)
     }
 
     #[no_mangle]