+++ example/lib/src/loopback.dart
@@ -119,7 +119,7 @@ class _LoopbackState extends State<Loopback> {
       // TODO: add some visualization via some progress indicator?
       var audioTrack =
           _tracks!.firstWhere((track) => track.kind() == MediaKind.audio);
-      audioTrack.onVolume((volume) => print("Volume update: $volume"));
+      audioTrack.onAudioLevelChanged((volume) => print("Volume update: $volume"));
 
       await vtrans?.sender.replaceTrack(
           _tracks!.firstWhere((track) => track.kind() == MediaKind.video));