+++ lib/src/platform/web/media_stream_track.dart
@@ -30,11 +30,11 @@ class WebMediaStreamTrack extends MediaStreamTrack {
   }
 
   @override
-  void onVolume(OnVolumeCallback? cb) {
+  void onAudioLevelChanged(OnAudioLevelChangedCallback? cb) {
     // TODO(review): is it? there is audio level in media-source stats and
     //               audio node and AnalyserNode
     //               https://webrtc.github.io/samples/src/content/getusermedia/volume/
-    throw 'onVolume callback is not supported on web';
+    throw 'onAudioLevelChanged callback is not supported on web';
   }
 
   @override